refactor: renamed components directory to onboarding
