+++ extension/source/QuillPage/Onboarding/LogoFooter.tsx
