+++ extension/source/QuillPage/Onboarding/OnboardingInfoPanel.tsx
