+++ extension/source/QuillPage/Onboarding/PasswordCreationPanel.tsx
