+++ extension/source/QuillPage/Onboarding/ReviewSecretPhrasePanel.tsx
