+++ extension/source/QuillPage/Onboarding/SecretPhrasePanel.tsx
