+++ extension/source/QuillPage/Onboarding/SetNicknamePanel.tsx
