+++ extension/source/QuillPage/Onboarding/ViewSecretPhrasePanel.tsx
