--- extension/source/QuillPage/components/QuickColumn.tsx
@@ -1,7 +1,0 @@
-import { FunctionComponent } from 'react';
-
-const QuickColumn: FunctionComponent = (props) => (
-  <div className="quick-column">{props.children}</div>
-);
-
-export default QuickColumn;
