--- extension/source/QuillPage/components/QuickRow.tsx
@@ -1,7 +1,0 @@
-import { FunctionComponent } from 'react';
-
-const QuickRow: FunctionComponent = (props) => (
-  <div className="quick-row">{props.children}</div>
-);
-
-export default QuickRow;
